Add http request related utilities
Includes a simple monadic wrapper for http requests, and applies that
to a specific safe error type so that no data is lost and everything
is propperly deserialized from all requests.

This monadic wrapper is used on all http verbs that we plan to use on
our api, being able to be extended with other verbs with a trivial
implementation.