Use shadowJar instead of includeApi
For whatever reason it doesn't work for me, and what it matters most is that it works. If its unsupported or, dare I say "legacy", I don't care.

If it compiles it's good. I'd commit crimes to have cargo on kotlin, I swear to god that gradle may have led people to insanity at best, suicide at worst.

And the worst part is that it's not the worst offender, java build tools can always get worse.